% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotypePhebruary2023.R
\docType{package}
\name{phenotypePhebruary2023-package}
\alias{phenotypePhebruary2023}
\alias{phenotypePhebruary2023-package}
\title{phenotypePhebruary2023: A Study Package for Generating Cohort Diagnostics}
\description{
A package that generates cohort diagnostics for cohort definitions included in the package.
}
\keyword{internal}
